analyzer/requirements.txt
fastapi
uvicorn
python-whois
aiofiles